Bug 929654 - Part 1: Prepare test activity setUp and tearDown for IDE launching. r=jmaher
This patch does two things.  First, it unifies the test activity
creation process, in preparation for future uses.  Second, it adds new
arguments -- "quit" and "finish" -- to the test activity destruction
process.  "quit" uses Robocop:Quit to kill Fennec and "finish" finishes
all activities started by test code.

Alternate test launchers (like from IDEs) don't have a good mechanism to
pass arguments, so the default must be suitable for such mechanisms, and
the best default for local testers is to not quit Fennec and to not
finish activities.  Therefore, these arguments default to "0" (off) and
are set to "1" (on) in runtestsremote.py; this preserves the existing
behaviour in automation.