Bug 929654 - Part 8: Implement --no-autorun for Robocop. r=jmaher
This adds an activity to the Robocop package that starts Fennec.  This
is needed because we pass configuration via robotium.config, so the
configuration needs to be interpreted by something running on device;
and it shouldn't be anything in Fennec itself.  This new activity starts
Fennec with the appropriate configuration.  The test harness then waits
forever.